User specific content, with JWT claims.

+++ VotjiAPI/Controllers/V1/PostsController.cs
@@ -11,6 +11,7 @@ using VotjiAPI.Contracts.V1.Responses;
 using VotjiAPI.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using VotjiAPI.Extensions;
 
 namespace VotjiAPI.Controllers.V1
 {
@@ -33,11 +34,15 @@ namespace VotjiAPI.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatePostRequest request)
         {
-            var post = new Post 
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if (!userOwnsPost)
             {
-                Id = postId,
-                Name = request.Name
-            };
+                return BadRequest(new { error = "You do not own this Post" });
+            }
+
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = request.Name;
+
             var updated = await _postService.UpdatePostAsync(post);
 
             if (updated)
@@ -49,6 +54,12 @@ namespace VotjiAPI.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own this Post" });
+            }
+
             var deleted = await _postService.DeletePostAsync(postId);
             if (deleted)
                 return NoContent();
@@ -69,10 +80,13 @@ namespace VotjiAPI.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequests)
         {
-            var post = new Post { Name = postRequests.Name };
-            
-
-                await _postService.CreatePostAsync(post);
+            var post = new Post
+            {
+                Name = postRequests.Name,
+                UserId = HttpContext.GetUserId()
+            };
+                
+            await _postService.CreatePostAsync(post);
 
             var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.ToUriComponent()}";
             var locationUrl = baseUrl + "/" + ApiRoutes.Posts.Get.Replace("{postId}", post.Id.ToString());
